2826: change clang warning settings and fix remaining warnings in code

+++ common/src/Renderer/EdgeRenderer.h
@@ -42,13 +42,13 @@ namespace TrenchBroom {
         public:
             struct Params {
                 float width;
-                float offset;
+                double offset;
                 bool onTop;
                 bool useColor;
                 Color color;
-                Params(float i_width, float i_offset, bool i_onTop);
-                Params(float i_width, float i_offset, bool i_onTop, const Color& i_color);
-                Params(float i_width, float i_offset, bool i_onTop, bool i_useColor, const Color& i_color);
+                Params(float i_width, double i_offset, bool i_onTop);
+                Params(float i_width, double i_offset, bool i_onTop, const Color& i_color);
+                Params(float i_width, double i_offset, bool i_onTop, bool i_useColor, const Color& i_color);
             };
 
             class RenderBase {
@@ -65,13 +65,13 @@ namespace TrenchBroom {
         public:
             virtual ~EdgeRenderer();
 
-            void render(RenderBatch& renderBatch, float width = 1.0f, float offset = 0.0f);
-            void render(RenderBatch& renderBatch, const Color& color, float width = 1.0f, float offset = 0.0f);
-            void render(RenderBatch& renderBatch, bool useColor, const Color& color, float width = 1.0f, float offset = 0.0f);
-            void renderOnTop(RenderBatch& renderBatch, float width = 1.0f, float offset = 0.2f);
-            void renderOnTop(RenderBatch& renderBatch, const Color& color, float width = 1.0f, float offset = 0.2f);
-            void renderOnTop(RenderBatch& renderBatch, bool useColor, const Color& color, float width = 1.0f, float offset = 0.2f);
-            void render(RenderBatch& renderBatch, bool useColor, const Color& color, bool onTop, float width, float offset);
+            void render(RenderBatch& renderBatch, float width = 1.0f, double offset = 0.0);
+            void render(RenderBatch& renderBatch, const Color& color, float width = 1.0f, double offset = 0.0);
+            void render(RenderBatch& renderBatch, bool useColor, const Color& color, float width = 1.0f, double offset = 0.0);
+            void renderOnTop(RenderBatch& renderBatch, float width = 1.0f, double offset = 0.2);
+            void renderOnTop(RenderBatch& renderBatch, const Color& color, float width = 1.0f, double offset = 0.2);
+            void renderOnTop(RenderBatch& renderBatch, bool useColor, const Color& color, float width = 1.0f, double offset = 0.2);
+            void render(RenderBatch& renderBatch, bool useColor, const Color& color, bool onTop, float width, double offset);
         private:
             virtual void doRender(RenderBatch& renderBatch, const Params& params) = 0;
         };
