+++ common/src/Renderer/RenderUtils.h
@@ -36,7 +36,7 @@ namespace TrenchBroom {
     namespace Renderer {
         class Vbo;
 
-        void glSetEdgeOffset(float f);
+        void glSetEdgeOffset(double f);
         void glResetEdgeOffset();
 
         void coordinateSystemVerticesX(const vm::bbox3f& bounds, vm::vec3f& start, vm::vec3f& end);
